work on migrating String+ConstString->std::string

+++ src/libYARP_OS/src/RosNameSpace.cpp
@@ -32,7 +32,7 @@ Contact RosNameSpace::queryName(const char *name) {
     ConstString full = name;
     ConstString node = full;
     ConstString srv = "";
-    int srv_idx = full.find("#");
+    size_t srv_idx = full.find("#");
     if (srv_idx!=ConstString::npos) {
         node = full.substr(0,srv_idx);
         srv = full.substr(srv_idx+1,full.length());
@@ -117,8 +117,8 @@ Contact RosNameSpace::registerName(const char *name) {
 Contact RosNameSpace::registerContact(const Contact& contact) {
     ConstString full = contact.getName();
     ConstString name = full;
-    int pub_idx = name.find("+#");
-    int sub_idx = name.find("-#");
+    size_t pub_idx = name.find("+#");
+    size_t sub_idx = name.find("-#");
 
     ConstString node = "";
     ConstString pub = "";
@@ -263,8 +263,8 @@ Contact RosNameSpace::registerContact(const Contact& contact) {
 Contact RosNameSpace::unregisterName(const char *rname) {
     ConstString full = rname;
     ConstString name = full;
-    int pub_idx = name.find("+#");
-    int sub_idx = name.find("-#");
+    size_t pub_idx = name.find("+#");
+    size_t sub_idx = name.find("-#");
 
     ConstString node = "";
     ConstString pub = "";
@@ -444,7 +444,7 @@ bool RosNameSpace::writeToNameServer(PortWriter& cmd,
 ConstString RosNameSpace::toRosName(const ConstString& name) {
     if (name.find(":")==ConstString::npos) return name;
     ConstString result;
-    for (int i=0; i<name.length(); i++) {
+    for (size_t i=0; i<name.length(); i++) {
         if (name[i]!=':') {
             result += name[i];
         } else {
@@ -459,7 +459,7 @@ ConstString RosNameSpace::fromRosName(const ConstString& name) {
     // length is at least 2
     ConstString result;
     int ct = 0;
-    for (int i=0; i<name.length(); i++) {
+    for (size_t i=0; i<name.length(); i++) {
         if (name[i]!='_') {
             if (ct) result += '_';
             result += name[i];
